requirements/base.txt
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
django-ckeditor==6.5.1
djangorestframework-simplejwt==5.2.2
django-mptt==0.14.0
django-environ==0.9.0
drf-yasg==1.21.4
psycopg2_binary==2.9.5
Pillow==9.3.0
six==1.16.0
Faker==16.6.1
faker-commerce==1.0.3
model-bakery==1.10.1
django-admin-hstore-widget==1.2.1
